backend/requirements.txt
Django==3.2.13
pymongo==3.12.1
environs==7.3.1
djongo==1.3.6
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
django-cors-headers==3.13.0
